splitted user/event folders, added auth
