+++ packages/server/package.json
@@ -20,6 +20,7 @@
     "graphql-subscriptions": "^1.1.0",
     "http": "^0.0.0",
     "isomorphic-fetch": "^2.2.1",
+    "jsonwebtoken": "^8.5.1",
     "kcors": "^2.2.2",
     "koa": "^2.8.1",
     "koa-bodyparser": "^4.2.1",
