+++ packages/server/src/auth.js
@@ -1,0 +1,22 @@
+import jwt from "jsonwebtoken";
+import { User } from "./modules/user/UserModel";
+
+export async function getUser(token) {
+  if (!token) return { user: null };
+
+  try {
+    const decodedToken = jwt.verify(token.substring(4), "secret_key");
+
+    const user = await User.findOne({ _id: decodedToken.id });
+
+    return {
+      user
+    };
+  } catch (err) {
+    return { user: null };
+  }
+}
+
+export function generateToken(user) {
+  return `JWT ${jwt.sign({ id: user._id }, jwtSecret)}`;
+}
