+++ packages/server/src/modules/user/mutation/index.js
@@ -1,0 +1,7 @@
+import UserCreate from "./UserRegisterWIthMutation";
+import UserLogin from "./UserLoginWIthMutation";
+
+export default {
+  UserCreate,
+  UserLogin
+};
